:hammer: move db directory
